refactor: move command registry to plumbing/
The register_command decorator and get_registry function are pure-stdlib
utilities with zero legion imports. Moving them from cli/registry.py to
plumbing/registry.py makes them accessible to other surfaces (cli_dev)
without cross-surface imports. See ADR-0008.
